change result value to double

+++ lesson05/app/src/main/java/com/github/rtyvz/senla/tr/lesson05/MainActivity.kt
@@ -21,7 +21,7 @@ class MainActivity : AppCompatActivity() {
                 if (firstValue.isBlank() || secondValue.isBlank()) {
                     result.text = getString(R.string.input_error)
                 } else {
-                    result.text = (firstValue.toInt() + secondValue.toInt()).toString()
+                    result.text = (firstValue.toDouble() + secondValue.toDouble()).toString()
                 }
             }
         }
